Patch 0073
・PTのアイテム分配実装 from eA（int_party.c, clif.c, intif.*, party.*, pc.c）
	完全にランダムで取得者を選定します
・battle_athena.confにparty_item_share_type, party_item_share_showを追加（battle.*）
・パーティ設定を変更するコマンド@partyoption追加（atcommand.*）
	通常クライアントからは変更できないアイテム収集方式・分配に関する設定を
	変更するためのコマンドで、PTリーダーのみ効果があります
・あまり意味がないので MIN_ATTRIBUTE, MAX_ATTRIBUTE, ATTRIBUTE_NORMAL, MIN_STAR, MAX_STAR
　のdefineを削除（mmo.h, atcommand.c）

+++ src/map/intif.c
@@ -436,7 +436,7 @@ void intif_party_addmember(struct map_session_data *sd)
 }
 
 // パーティ設定変更
-void intif_party_changeoption(int party_id, int account_id, unsigned short exp, unsigned short item)
+void intif_party_changeoption(int party_id, int account_id, int exp, int item)
 {
 	if (inter_fd < 0)
 		return;
