+++ src/map/intif.h
@@ -21,7 +21,7 @@ int intif_unlock_guild_storage(int guild_id);
 void intif_create_party(struct map_session_data *sd, char *name, int item, int item2);
 int intif_request_partyinfo(int party_id);
 void intif_party_addmember(struct map_session_data *sd);
-void intif_party_changeoption(int party_id, int account_id, unsigned short exp, unsigned short item);
+void intif_party_changeoption(int party_id, int account_id, int exp, int item);
 void intif_party_leave(int party_id, int account_id, const char * name);
 void intif_party_changemap(struct map_session_data *sd, unsigned char online);
 int intif_break_party(int party_id);
